Cleaned up test structure and fixed test discovery
