+++ tests/test_calibration/test_vis_calibration_apply.py
@@ -7,10 +7,6 @@ from PyFHD.use_idl_fhd.use_idl_outputs import convert_sav_to_dict
 from PyFHD.pyfhd_tools.test_utils import recarray_to_dict, sav_file_vis_arr_swap_axes
 import numpy as np
 import deepdish as dd
-import matplotlib.pyplot as plt
-import importlib_resources
-
-from logging import RootLogger
 
 @pytest.fixture
 def data_dir():
